9077: trying something else with an isolated journey test to just add practitioners to the contactPrimary [skip ci]
