+++ web-client/integration-tests/journey/petitionsClerkAddsPractitionersToCase.js
@@ -1,4 +1,5 @@
 import { formattedCaseDetail as formattedCaseDetailComputed } from '../../src/presenter/computeds/formattedCaseDetail';
+import { petitionsClerkAddsPractitionerToPrimaryContact } from './petitionsClerkAddsPractitionerToPrimaryContact';
 import { refreshElasticsearchIndex } from '../helpers';
 import { runCompute } from 'cerebral/test';
 import { withAppContextDecorator } from '../../src/withAppContext';
@@ -23,13 +24,13 @@ export const petitionsClerkAddsPractitionersToCase = (
       [],
     );
 
-    await associatePrivatePractitionerToPrimaryPetitioner(
+    await petitionsClerkAddsPractitionerToPrimaryContact(
       cerebralTest,
       practitionerBarNumber,
     );
 
     if (addSecondaryPractitionerToPetitioner) {
-      await associatePrivatePractitionerToPrimaryPetitioner(
+      await petitionsClerkAddsPractitionerToPrimaryContact(
         cerebralTest,
         'PT5432',
       );
@@ -88,86 +89,4 @@ export const petitionsClerkAddsPractitionersToCase = (
       await refreshElasticsearchIndex();
     }
   });
-};
-
-const associatePrivatePractitionerToPrimaryPetitioner = async (
-  cerebralTest,
-  barNumber,
-) => {
-  await cerebralTest.runSequence('openAddPrivatePractitionerModalSequence');
-
-  expect(
-    cerebralTest.getState('validationErrors.practitionerSearchError'),
-  ).toBeDefined();
-
-  await cerebralTest.runSequence('updateFormValueSequence', {
-    key: 'practitionerSearch',
-    value: barNumber,
-  });
-
-  await cerebralTest.runSequence('openAddPrivatePractitionerModalSequence');
-
-  expect(
-    cerebralTest.getState('validationErrors.practitionerSearchError'),
-  ).toBeUndefined();
-  expect(cerebralTest.getState('modal.practitionerMatches.length')).toEqual(1);
-
-  //default selected because there was only 1 match
-  let practitionerMatch = cerebralTest.getState('modal.practitionerMatches.0');
-  expect(cerebralTest.getState('modal.user.userId')).toEqual(
-    practitionerMatch.userId,
-  );
-
-  const formattedCase = runCompute(formattedCaseDetail, {
-    state: cerebralTest.getState(),
-  });
-  const contactPrimary = formattedCase.petitioners[0];
-
-  await cerebralTest.runSequence('updateModalValueSequence', {
-    key: `representingMap.${contactPrimary.contactId}`,
-    value: true,
-  });
-
-  if (cerebralTest.intervenorContactId) {
-    await cerebralTest.runSequence('updateModalValueSequence', {
-      key: `representingMap.${cerebralTest.intervenorContactId}`,
-      value: true,
-    });
-  }
-
-  expect(
-    cerebralTest.getState('validationErrors.practitionerSearchError'),
-  ).toBeUndefined();
-
-  await cerebralTest.runSequence(
-    'associatePrivatePractitionerWithCaseSequence',
-  );
-
-  expect(
-    cerebralTest.getState('caseDetail.privatePractitioners.length'),
-  ).toEqual(1);
-  expect(
-    cerebralTest.getState('caseDetail.privatePractitioners.0.representing'),
-  ).toContain(contactPrimary.contactId);
-
-  if (cerebralTest.intervenorContactId) {
-    expect(
-      cerebralTest.getState('caseDetail.privatePractitioners.0.representing'),
-    ).toContain(cerebralTest.intervenorContactId);
-  }
-
-  expect(
-    cerebralTest.getState('caseDetail.privatePractitioners.0.name'),
-  ).toEqual(practitionerMatch.name);
-
-  let formatted = runCompute(formattedCaseDetail, {
-    state: cerebralTest.getState(),
-  });
-
-  expect(formatted.privatePractitioners.length).toEqual(1);
-  expect(formatted.privatePractitioners[0].formattedName).toEqual(
-    `${practitionerMatch.name} (${practitionerMatch.barNumber})`,
-  );
-
-  cerebralTest.privatePractitioners = formatted.privatePractitioners[0];
 };
